requirements.txt
pytest
requests 
mysql-connector-python 
python-dotenv 